Refactor for Security
