+++ module-api/src/main/java/com/myfin/api/service/impl/AccountServiceImpl.java
@@ -1,7 +1,7 @@
 package com.myfin.api.service.impl;
 
 import com.myfin.api.dto.CreateAccount;
-import com.myfin.api.service.ATopServiceComponent;
+import com.myfin.api.service.TopServiceComponent;
 import com.myfin.api.service.AccountService;
 import com.myfin.core.dto.AccountDto;
 import lombok.RequiredArgsConstructor;
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class AccountServiceImpl extends ATopServiceComponent implements AccountService {
+public class AccountServiceImpl extends TopServiceComponent implements AccountService {
 
     @Override
     public AccountDto createAccount(CreateAccount.Request request) {
