+++ module-api/src/main/java/com/myfin/api/service/impl/UserAuthenticationServiceImpl.java
@@ -1,0 +1,32 @@
+package com.myfin.api.service.impl;
+
+import com.myfin.api.service.UserAuthenticationService;
+import com.myfin.core.exception.impl.NotFoundException;
+import com.myfin.core.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class UserAuthenticationServiceImpl implements UserAuthenticationService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return userRepository.findById(username)
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 유저입니다"));
+    }
+
+    @Override
+    public Authentication getAuthentication(String userId) {
+        UserDetails userDetails = this.loadUserByUsername(userId);
+        return new UsernamePasswordAuthenticationToken(userDetails, userDetails.getPassword(), userDetails.getAuthorities());
+    }
+}
