+++ module-api/src/main/java/com/myfin/api/service/impl/UserSignUpServiceImpl.java
@@ -4,7 +4,7 @@ import com.myfin.adapter.coolsms.SMSMessageComponent;
 import com.myfin.api.dto.SignUp;
 import com.myfin.api.dto.VerifyIdentity;
 import com.myfin.api.dto.VerifyIdentityResultDto;
-import com.myfin.api.service.ATopServiceComponent;
+import com.myfin.api.service.TopServiceComponent;
 import com.myfin.api.service.UserSignUpService;
 import com.myfin.cache.entity.CacheVerified;
 import com.myfin.cache.entity.CacheVerifyCode;
@@ -29,7 +29,7 @@ import java.time.LocalDateTime;
 @Service
 @Slf4j
 @RequiredArgsConstructor
-public class UserSignUpServiceImpl extends ATopServiceComponent implements UserSignUpService {
+public class UserSignUpServiceImpl extends TopServiceComponent implements UserSignUpService {
 
     private final UserRepository userRepository;
     private final CacheVerifyCodeRepository cacheVerifyCodeRepository;
