+++ module-security/build.gradle
@@ -16,9 +16,6 @@ repositories {
 }
 
 dependencies {
-//    implementation project(':module-core')
-
-    implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'io.jsonwebtoken:jjwt:0.9.1'
     implementation 'javax.xml.bind:jaxb-api:2.3.1'
