--- module-security/src/main/java/com/myfin/security/service/UserAuthenticationComponent.java
@@ -1,17 +1,0 @@
-package com.myfin.security.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class UserAuthenticationComponent implements UserDetailsService {
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return null;
-    }
-}
